Restructure redux folder and file setup
