+++ src/redux/store.ts
@@ -1,6 +1,6 @@
 // src/store/store.ts
 import { configureStore } from '@reduxjs/toolkit';
-import authReducer from './authSlice';
+import authReducer from './features/authSlice';
 
 export const store = configureStore({
   reducer: {
